% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{add_tally}
\alias{add_tally}
\title{Add count of observations to corpus}
\usage{
add_tally(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{additional arguments passed to methods}
}
\description{
\code{add_tally} is a generic function for adding a count column. The default
method calls \code{\link[dplyr:count]{dplyr::add_tally()}}.
}
