% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_rows.R
\name{bind_rows}
\alias{bind_rows}
\alias{bind_rows.corpus}
\title{Bind multiple corpora by row}
\usage{
\method{bind_rows}{corpus}(..., .id = NULL, .meta = NULL)
}
\arguments{
\item{...}{comma-separated list of corpus objects}

\item{.id}{corpus identifier. When \code{.id} is supplied, a document variable of identifiers is
created to link each document to its corpus.}

\item{.meta}{a named list of metadata fields in the corpus. If \code{.meta} is \code{NULL},
the default, the metadata of the first corpus supplied to \code{bind_rows} will be used.}
}
\description{
\code{bind_rows()} appends documents and document variables
}
\examples{
corp1 <- data_corpus_inaugural[1:5]
corp2 <- data_corpus_inaugural[6:10]

# using the metadata of the first corpus by default
bind_rows(corp1, corp2)

# supply custom meta data 
md <- list(source = "some source", author = "some author")
bind_rows(corp1, corp2, .meta = md)

# create document variable named "corpus_id" 
bind_rows(corp1, corp2, .id = "corpus_id")
}
