% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler_functions.R
\name{dplyr-wrappers}
\alias{dplyr-wrappers}
\alias{corpus_stv_byvar}
\alias{corpus_stv_bydoc}
\title{Wrappers to dplyr functions}
\usage{
corpus_stv_byvar(.data, ..., fun)

corpus_stv_bydoc(.data, ..., fun)
}
\arguments{
\item{.data}{input quanteda object}

\item{...}{arguments for the dplyr function}

\item{fun}{reference to the dplyr function}
}
\value{
a modified \pkg{quanteda} object
}
\description{
Wrapper functions for \pkg{dplyr} functions to preserve texts, document
names, and corpus meta-data.
}
\keyword{internal}
