% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R, R/dplyr_groups.R
\name{add_count}
\alias{add_count}
\alias{dplyr_groups}
\title{Verbs that operate on groups of rows}
\arguments{
\item{x}{a \pkg{quanteda} corpus object}

\item{...}{additional arguments passed to methods}
}
\value{
A corpus with an additional document variable containing counts.
}
\description{
These functions operate on groups of rows (documents) of \pkg{quanteda}
objects, typically counting or summarising documents by group.
}
\details{
\code{add_count()} and \code{add_tally()} add a document variable containing the count
of observations in each group. See \code{\link[dplyr:count]{dplyr::add_count()}} for more details.
}
\keyword{internal}
