% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R, R/dplyr_single_table_fns.R,
%   R/filter.R, R/mutate.R, R/pull.R, R/select.R, R/slice.R
\name{arrange}
\alias{arrange}
\alias{dplyr_single}
\alias{filter}
\alias{mutate}
\alias{transmute}
\alias{pull}
\alias{select}
\alias{rename}
\alias{slice}
\alias{slice_head}
\alias{slice_tail}
\alias{slice_sample}
\alias{slice_min}
\alias{slice_max}
\title{Single-table verbs from dplyr}
\description{
These are data manipulation functions designed to work on \pkg{quanteda}
objects, where the "single table" is the docvars from the object.
Each function operates on the docvars and returns a modified version of the
original object.
}
\details{
\code{arrange} sorts documents by values of the docvars.

\code{filter} subsets documents. See \code{\link[dplyr:filter]{dplyr::filter()}}
for more details.

\code{mutate} and \code{transmute} can add or modify document variables. See
\code{\link[dplyr:mutate]{dplyr::mutate()}} for more details.

\code{pull} extracts a single docvar as a vector.
See \code{\link[dplyr:pull]{dplyr::pull()}} for more details.

\code{select} and \code{rename} keep or rename document variables. See
\code{\link[dplyr:select]{dplyr::select()}} for more details.

\code{slice} lets you index documents by their (integer) locations, to select,
remove, and duplicate documents.  See \code{\link[dplyr:slice]{dplyr::slice()}} for more details.
}
\keyword{internal}
