% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_vector_fns.R
\name{dplyr_vector}
\alias{dplyr_vector}
\alias{desc}
\title{Vector functions from dplyr}
\arguments{
\item{x}{a vector to transform}
}
\value{
A transformed vector of the same length as the input.
}
\description{
Selected vector functions, re-exported.

\code{desc()} reverses the sort order of a variable; see
\code{\link[dplyr:desc]{dplyr::desc()}} for details.
}
\keyword{internal}
