% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.textstat}
\alias{filter.textstat}
\title{Return rows with matching conditions including feature matches}
\usage{
\method{filter}{textstat}(
  .data,
  ...,
  pattern = NULL,
  valuetype = c("glob", "regex", "fixed"),
  case_insensitive = TRUE
)
}
\arguments{
\item{.data}{a \code{textstat} object returned from one of the applicable
\code{textstat_*} functions in \pkg{quanteda}}

\item{...}{filter conditions passed to \pkg{dplyr} \code{\link[dplyr]{filter}}}

\item{pattern}{a character vector, list of character vectors,
\link[quanteda]{dictionary}, or \link[quanteda]{collocations} object.  See \link[quanteda]{pattern} for
details.}

\item{valuetype}{the type of pattern matching: \code{"glob"} for "glob"-style
wildcard expressions; \code{"regex"} for regular expressions; or \code{"fixed"} for
exact matching. See \link[quanteda]{valuetype} for details.}

\item{case_insensitive}{ignore case when matching, if \code{TRUE}}
}
\description{
Use filters to find rows of the return objects from many \code{textstat_*()}
functions where conditions are true, including matches to features using
\code{"glob"}, \code{"regex"} or \code{"fixed"} patterns.
}
\examples{
period <- ifelse(docvars(data_corpus_inaugural, "Year") < 1945, "pre-war", "post-war")
mydfm <- dfm(data_corpus_inaugural, groups = period)
keyness <- textstat_keyness(mydfm)
filter(keyness, pattern = "america*")
filter(keyness, p < .00001, pattern = "america*") 
filter(keyness, p < .00001) \%>\% head()
}
\seealso{
\code{\link[quanteda]{textstat_collocations}},
\code{\link[quanteda]{textstat_keyness}},
\code{\link[quanteda]{textstat_frequency}}
}
\keyword{internal}
