% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse.R
\name{glimpse}
\alias{glimpse}
\alias{glimpse.corpus}
\title{Get a glimpse of a quanteda object}
\usage{
\method{glimpse}{corpus}(x, width = NULL, ...)
}
\arguments{
\item{x}{a corpus or \pkg{quanteda} object}

\item{width}{width of the output; default to the width of the console}

\item{...}{unused}
}
\value{
Invisibly returns the input corpus. Called primarily for its
side effect of printing a summary to the console.
}
\description{
Implementation of \link[tibble:reexports]{glimpse} for \pkg{quanteda} objects,
allowing docvars to be viewed.
}
\examples{
glimpse(data_corpus_inaugural)
}
\keyword{internal}
