% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda.tidy-package.R
\docType{package}
\name{quanteda.tidy-package}
\alias{quanteda.tidy}
\alias{quanteda.tidy-package}
\title{quanteda.tidy: 'tidyverse' Extensions for 'quanteda'}
\description{
Adds tidy functionality, especially 'dplyr' functions, to 'quanteda' objects.
}
\author{
\strong{Maintainer}: Kenneth Benoit \email{kbenoit@smu.edu.sg} [copyright holder]

}
\keyword{internal}
