% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.corpus}
\alias{select.corpus}
\alias{rename.corpus}
\title{Select/rename docvars by name}
\usage{
\method{select}{corpus}(.data, ...)

\method{rename}{corpus}(.data, ...)
}
\arguments{
\item{.data}{a \pkg{quanteda} object with document variables}

\item{...}{comma-separated list of unquoted document variables.  See
\link[dplyr:select]{select}.}
}
\description{
Choose or rename document variables.  \code{select()} keeps only the docvars you
mention; \code{rename()} keeps all of the docvars.
}
\examples{
data_corpus_inaugural \%>\%
  select(Party, Year) \%>\%
  summary(n = 5)

data_corpus_inaugural \%>\%
  rename(LastName = President) \%>\%
  select(FirstName, LastName) \%>\%
  summary(n = 5)
}
